Changes to enable browser-specific authentication functionality.
