+++ lib/auth/iam/auth_provider.js
@@ -25,7 +25,10 @@ const InstancePrincipalProvider = require('./instance_principal');
 /* Maximum lifetime of signature 300 seconds */
 const MAX_ENTRY_LIFE_TIME = 300;
 
-const SIGNING_HEADERS = '(request-target) host date';
+const DATE_HEADER = Utils.isInBrowser ?
+    HttpConstants.X_DATE : HttpConstants.DATE;
+
+const SIGNING_HEADERS = `(request-target) host ${DATE_HEADER}`;
 
 const OBO_TOKEN_HEADER = 'opc-obo-token';
 
@@ -108,6 +111,12 @@ credentials provider properties together with user identity properties', cfg);
 properties together with credentials provider', cfg);
             }
             this._provider = new UserProfileProvider(cfg);
+        } else if (opt.profileProvider != null) {
+            if (typeof opt.profileProvider !== 'object') {
+                throw new NoSQLArgumentError(
+                    'Custom profile provider must be an object', cfg);
+            }
+            this._provider = opt.profileProvider;
         } else {
             this._provider = new OCIConfigFileProvider(cfg);
         }
@@ -209,7 +218,7 @@ is invalid or empty');
         let content = `${HttpConstants.REQUEST_TARGET}: post /\
 ${HttpConstants.NOSQL_DATA_PATH}\n\
 ${HttpConstants.HOST}: ${this._serviceHost}\n\
-${HttpConstants.DATE}: ${dateStr}`;
+${DATE_HEADER}: ${dateStr}`;
         if (this._delegationToken != null) {
             content += `\n${OBO_TOKEN_HEADER}: ${this._delegationToken}`;
         }
@@ -223,7 +232,7 @@ ${HttpConstants.DATE}: ${dateStr}`;
         }
         const date = new Date();
         const dateStr = date.toUTCString();
-        let signature = Utils.sign(this._signingContent(dateStr),
+        let signature = await Utils.sign(this._signingContent(dateStr),
             this._profile.privateKey, 'request');
         return {
             time: date.getTime(),
@@ -288,7 +297,7 @@ ${HttpConstants.DATE}: ${dateStr}`;
         }
         const ret = {
             [HttpConstants.AUTHORIZATION]: this._signatureDetails.header,
-            [HttpConstants.DATE]: this._signatureDetails.dateStr,
+            [DATE_HEADER]: this._signatureDetails.dateStr,
         };
 
         //It is possible that if _createSignatureDetails() is called
@@ -315,6 +324,15 @@ ${HttpConstants.DATE}: ${dateStr}`;
         if (compartment != null) {
             ret[HttpConstants.COMPARTMENT_ID] = compartment;
         }
+
+        //Currently proxy uses presence this header to identify requests from
+        //the browser and thus enable CORS (by sending back
+        //Access-Control-Allow-Origin header).  The value of this header is
+        //not currently used.  
+        if (Utils.isInBrowser) {
+            ret[HttpConstants.OPC_REQUEST_ID] = 1;
+        }
+
         return ret;
     }
