+++ lib/auth/iam/instance_principal.js
@@ -74,8 +74,8 @@ ${HttpConstants.APPLICATION_JSON}\n\
 ${HttpConstants.CONTENT_SHA256}: ${digest}`;
     }
 
-    _authHeader(instCert, instPrivateKey, dateStr, payload, digest) {
-        const signature = Utils.sign(
+    async _authHeader(instCert, instPrivateKey, dateStr, payload, digest) {
+        const signature = await Utils.sign(
             this._signingContent(dateStr, payload, digest),
             instPrivateKey,
             'instance principal federation request');
@@ -84,10 +84,10 @@ ${HttpConstants.CONTENT_SHA256}: ${digest}`;
         return Utils.signatureHeader(SIGNING_HEADERS, keyId, signature);
     }
 
-    _requestHeaders(instCert, instPrivateKey, payload) {
+    async _requestHeaders(instCert, instPrivateKey, payload) {
         const dateStr = new Date().toUTCString();
         const digest = Utils.sha256digest(payload);
-        const auth = this._authHeader(instCert, instPrivateKey, dateStr,
+        const auth = await this._authHeader(instCert, instPrivateKey, dateStr,
             payload, digest);
         return {
             [HttpConstants.CONTENT_TYPE]: HttpConstants.APPLICATION_JSON,
@@ -101,11 +101,13 @@ ${HttpConstants.CONTENT_SHA256}: ${digest}`;
         intermediateCerts) {
         const payload = this._requestPayload(publicKey, instCert,
             intermediateCerts);
-        
+        const requestHeaders = await this._requestHeaders(instCert,
+            instPrivateKey, payload);
+
         const req = {
             url: this._url,
             method: HttpConstants.POST,
-            headers: this._requestHeaders(instCert, instPrivateKey, payload),
+            headers: requestHeaders,
             timeout: this._timeout,
             payload
         };
