+++ lib/constants.js
@@ -510,7 +510,9 @@ const HttpConstants = {
     SET_COOKIE: 'set-cookie',
     HOST: 'host',
     DATE: 'date',
+    X_DATE: 'x-date',
     REQUEST_ID: 'x-nosql-request-id',
+    OPC_REQUEST_ID: 'opc-request-id',
     DATA_PATH_NAME: 'data',
     NOSQL_VERSION: 'V2',
     NOSQL_PATH_NAME: 'nosql',
